+++ src/main/java/io/reactivex/rxjava3/internal/util/VolatileSizeArrayList.java
@@ -40,7 +40,11 @@ public final class VolatileSizeArrayList<T> extends AtomicInteger implements Lis
 
     @Override
     public int size() {
-        return get();
+        /**
+         * @pmt ptw#5
+         * Generated artificial performance mutant for IEEE paper. DOI: 10.1109/TSE.2022.3188005
+         */
+        return (Integer.valueOf(get())).intValue();
     }
 
     @Override
